requirements.txt
bcrypt==4.2.1
bleach==6.2.0
cffi~=1.17.1
configparser==7.1.0
flask==3.1.0
Flask-Caching==2.3.0
flask-socketio==5.5.1
gunicorn==23.0.0
jinja2==3.1.6
markdown==3.7
mysql-connector-python==9.2.0
mysql==0.0.3
packaging==24.2
pillow==11.1.0
portalocker==3.1.1
pymysql==1.1.1
qrcode~=8.0
requests~=2.32.3
waitress~=3.0.2
Werkzeug~=3.1.3
user-agents==2.2.0
PyJWT==2.10.1
opencv-python~=4.11.0.86